Remove FileUpload in static mode
