+++ packages/bodiless-components/src/FileUpload/FileUpload.static.ts
@@ -1,0 +1,1 @@
+export { Fragment as FileUpload } from '@bodiless/fclasses';
